requirements.txt
flask
python-dotenv
requests
python-dateutil

